server/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
requests
python-dotenv
google-generativeai
pillow
